Rename: Database -> UserRepository로 이름 변경
앞으로 게시판 저장소도 만들어야함
