+++ src/main/java/controller/UserController.java
@@ -1,7 +1,7 @@
 package controller;
 
 import annotation.RequestMapping;
-import db.Database;
+import db.UserRepository;
 import http.HttpRequest;
 import http.HttpResponse;
 import model.User;
@@ -11,13 +11,13 @@ import java.util.UUID;
 
 public class UserController implements Controller{
 
-    private final Database database = Database.getInstance();
+    private final UserRepository userRepository = UserRepository.getInstance();
 
     @RequestMapping(value="/user/create", method = "POST")
     public String createUser(HttpRequest request, HttpResponse response) {
         Map<String, String> data = request.getFormData();
         User user = new User(data.get("userId"), data.get("password"), data.get("name"), data.get("email"));
-        database.addUser(user);
+        userRepository.addUser(user);
 
         return "redirect:/index";
     }
@@ -25,12 +25,12 @@ public class UserController implements Controller{
     @RequestMapping(value="/user/login", method = "POST")
     public String login(HttpRequest request, HttpResponse response) {
         Map<String, String> data = request.getFormData();
-        User user = database.findUserById(data.get("userId"));
+        User user = userRepository.findUserById(data.get("userId"));
 
         if (data.get("password").equals(user.getPassword())) {
             String sid = generateSessionId();
             response.setHeader("Set-Cookie","sid="+sid+"; Max-Age=300; Path=/");
-            SessionManager.addSession(sid, user);
+            SessionManager.addSession(sid, user.getUserId());
             return "redirect:/index";
         } else {
             return "redirect:/user/login_failed";
